requirements.txt
dvc==3.67.1
mlflow>=2.18.0
scikit-learn>=1.3.0
pandas>=2.0.0
pyyaml>=6.0
matplotlib>=3.7.0
